+++ tests/test_datasets/test_dataset.py
@@ -343,7 +343,7 @@ def test_create_ts_with_int_timestamp():
 
 
 @pytest.mark.filterwarnings(
-    "ignore: Timestamp contains numeric values, which can lead to unexpected results if freq is not None."
+    "ignore: Timestamp contains numeric values, and given freq is D. Timestamp will be converted to datetime."
 )
 def test_create_ts_with_int_timestamp_with_freq():
     df = generate_ar_df(periods=10, freq=None, n_segments=3)
@@ -386,7 +386,7 @@ def test_create_ts_with_exog_int_timestamp():
 
 
 @pytest.mark.filterwarnings(
-    "ignore: Timestamp contains numeric values, which can lead to unexpected results if freq is not None."
+    "ignore: Timestamp contains numeric values, and given freq is D. Timestamp will be converted to datetime."
 )
 def test_create_ts_with_exog_int_timestamp_with_freq():
     df = generate_ar_df(periods=10, start_time=5, freq=None, n_segments=3, random_seed=0)